+++ sentry-ruby/lib/sentry/configuration.rb
@@ -160,6 +160,10 @@ module Sentry
     # ```
     attr_accessor :traces_sampler
 
+    # Send diagnostic client reports about dropped events, true by default
+    # tries to attach to an existing envelope max once every 30s
+    attr_accessor :send_client_reports
+
     # these are not config options
     attr_reader :errors, :gem_specs
 
@@ -212,6 +216,7 @@ module Sentry
       self.send_modules = true
       self.send_default_pii = false
       self.skip_rake_integration = false
+      self.send_client_reports = true
       self.trusted_proxies = []
       self.dsn = ENV['SENTRY_DSN']
       self.server_name = server_name_from_env