+++ sentry-ruby/spec/sentry/client/event_sending_spec.rb
@@ -35,6 +35,7 @@ RSpec.describe Sentry::Client do
         it "doesn't send the event when it's not sampled" do
           allow(Random).to receive(:rand).and_return(0.51)
           subject.capture_event(event, scope)
+          expect(subject.transport).to have_recorded_lost_event(:sample_rate, 'event')
           expect(subject.transport.events.count).to eq(0)
         end
       end
@@ -156,6 +157,17 @@ RSpec.describe Sentry::Client do
           expect(transport.events.count).to eq(1)
         end
       end
+
+      it "records queue overflow" do
+        allow(Sentry.background_worker).to receive(:perform).and_return(false)
+
+        subject.capture_event(event, scope)
+        expect(subject.transport).to have_recorded_lost_event(:queue_overflow, 'event')
+
+        expect(transport.events.count).to eq(0)
+        sleep(0.2)
+        expect(transport.events.count).to eq(0)
+      end
     end
   end
 
@@ -269,6 +281,7 @@ RSpec.describe Sentry::Client do
         it "discards the event and logs a info" do
           expect(subject.capture_event(event, scope)).to be_nil
 
+          expect(subject.transport).to have_recorded_lost_event(:event_processor, 'event')
           expect(string_io.string).to match(/Discarded event because one of the event processors returned nil/)
         end
       end
@@ -309,6 +322,7 @@ RSpec.describe Sentry::Client do
         it "swallows and logs Sentry::ExternalError (caused by transport's networking error)" do
           expect(subject.capture_event(event, scope)).to be_nil
 
+          expect(subject.transport).to have_recorded_lost_event(:network_error, 'event')
           expect(string_io.string).to match(/Event sending failed: Failed to open TCP connection/)
           expect(string_io.string).to match(/Unreported Event: Test message/)
           expect(string_io.string).to match(/Event capturing failed: Failed to open TCP connection/)
@@ -333,6 +347,7 @@ RSpec.describe Sentry::Client do
           expect(subject.capture_event(event, scope)).to be_a(Sentry::Event)
           sleep(0.2)
 
+          expect(subject.transport).to have_recorded_lost_event(:network_error, 'event')
           expect(string_io.string).to match(/Event sending failed: Failed to open TCP connection/)
           expect(string_io.string).to match(/Unreported Event: Test message/)
         end
@@ -416,6 +431,19 @@ RSpec.describe Sentry::Client do
             expect(string_io.string).to match(/Event sending failed: TypeError/)
             expect(string_io.string).to match(__FILE__)
           end
+        end
+      end
+
+      context "before_send returns nil" do
+        before do
+          configuration.before_send = lambda do |_event, _hint|
+            nil
+          end
+        end
+
+        it "records lost event" do
+          subject.send_event(event)
+          expect(subject.transport).to have_recorded_lost_event(:before_send, 'event')
         end
       end
     end